fix(d19-m2a): resolve both review rounds — seal the study script, honesty-harden diagnostics, bind guards and tests (D20 review record)
codex gpt-5.6-sol (xhigh, read-only) on the PR diff: FIX-FIRST, 8 findings;
parallel 3-lens verification workflow (21 agents, adversarial refutation):
18 raw findings, 13 confirmed (5 overlapping). Every confirmed finding fixed:

- HIGH: bms_star_mauna_loa.py now consumes load_mauna_loa_training,
  evaluates on the training span only, and plots no sealed test value;
  source-level seal-guard test added (§6.6 made mechanical for the one
  study-facing script, matching what the prereg addendum claims).
- Diagnostics honesty: no-warmup leapfrog counts and partial-chain
  divergence/acceptance payloads now report as UNAVAILABLE instead of
  contaminated/fabricated values; acceptance validated finite in [0,1];
  to_json(allow_nan=False).
- A10: build_mauna_loa_kernels stamps _a10_frozen_period; fit_map asserts
  the stamped value at ENTRY (frozen-but-premutated periods no longer
  survive); hostile-optimizer test covers the exit guard's failing path.
- A4: CandidateResult carries universe identity; assert_single_universe
  validates the exact run_bms_star input; MAUNA_METRICS gains the
  DISTRIBUTION-LEVEL kl_forward the plan's "plus kl_forward" clause names
  (pw_kl_forward is one of the legacy five); script wiring source-pinned.
- ExponentialHarmonic2Model fallback sigma from full trend+harmonics
  residuals (was trend-only, inflating predictive variance).
- Canonical hash serialization explicitly little-endian ("<f8");
  public-loader gate tests + network-free openml-path tests added.
- Slurm guard recognizes python/python3/paths and interpreter options,
  with per-file coverage assertion.
- Vendored CSV shipped in wheels via [tool.setuptools.package-data].

Gates re-run on the final state: suite 170 passed (121 baseline + 49 new);
d19_prior_scorecard.py regenerated scorecard_v2.json byte-identically
(sha256 52b2d49d...6881). Full finding-by-finding record: DECISIONS.md D20
("M2a PR-review round" + "Parallel verification workflow").

+++ bistar_gp/data.py
@@ -113,7 +113,11 @@ def _verify_mauna_provenance(df):
             "(prereg §6.2 requires re-pinning before any use of deviating data)")
 
     def col_bytes(col):
-        return np.ascontiguousarray(df[col].astype(float).values).tobytes()
+        # Explicit little-endian float64 ("<f8"): the addendum freezes the
+        # byte serialization, and native-endian bytes would hard-fail a
+        # correct artifact on a big-endian host (M2a review round, finding 7).
+        return np.ascontiguousarray(
+            df[col].astype(float).values, dtype="<f8").tobytes()
 
     canonical = hashlib.sha256(
         col_bytes(year_col) + col_bytes(month_col) + col_bytes(co2_col)).hexdigest()
